Add etcd role
etcd is added in the project as a key-value storage and a service
discovery mechanism for Calico. etcd is used to distribute information
about workloads, Calico endpoints and policies.

It's started in a Docker container and managed by systemd on each host
for redundancy. Since we register etcd as a service in Consul, we could
use it for other purposes as well. Calico doesn't support other key-value
storages such as Consul, but we might want to extend it with such functionality
in the future.