Add Calico role
Calico is a new virtual network solution, and it's added in the project
to add the IP per container functionality. Calico connects Docker containers
via IP no matter which worker node they are on.

Calico uses etcd as a service discovery mechanism and a key-value storage
to distribute information about workloads. Any workload managed by Calico is
registered as a service in Consul as well. Thus, you have the option to
query Consul DNS in order to resolve IP addresses from Calico network.
Calico is deployed in a Docker container on each worker node and managed
by systemd.

All major cloud providers are supported such as AWS, DigitalOcean and GCE
by leveraging the IP in IP mode. The full node-to-node BGP mesh is supported
in Cisco OpenStack cloud only.